pyproject.toml
[project]
name = "meethub-backend"
version = "0.1.0"
description = "TikTok-like MVP backend on FastAPI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.30",
    "psycopg[binary]>=3.1.19",
    "pydantic-settings>=2.4.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.1.3",
    "redis>=5.0.7",
    "boto3>=1.34.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "email-validator>=2.2.0",
]

[tool.uv]
package = false
